Django_Project/requirements.txt
asgiref==3.11.0
crispy-bootstrap5==2025.6
Django==6.0.1
django-crispy-forms==2.5
django-debug-toolbar==6.2.0
django-environ==0.12.0
djangorestframework==3.16.1
pillow==12.1.1
psycopg2-binary==2.9.11
sqlparse==0.5.5
